backend/requirements.txt
# CrewAI Multi-Agent Framework (Python 3.11 recommended)
crewai
crewai-tools

# FastAPI and Core Dependencies
fastapi
uvicorn[standard]
pydantic>=2.0
pydantic-settings
python-multipart

# AI and LLM Dependencies (Groq only - no OpenAI)
langchain
langchain-groq
langchain-community
langchain-core
groq

# LangGraph for agent workflows
langgraph

# Audio and Speech Processing
SpeechRecognition
gTTS
pydub
pygame

# File and System Operations
watchdog
python-magic-bin
pathlib2
send2trash

# Utilities
python-dotenv
requests
aiofiles
websockets

# Development and Testing
pytest
pytest-asyncio
httpx

# Security and CORS
python-jose[cryptography]
passlib[bcrypt]

# MongoDB and Database
pymongo
motor

# Enhanced TTS and Conversational AI (Best Free Libraries)
edge-tts
pyttsx3
TTS
openai-whisper
vosk
soundfile
sounddevice
numpy
scipy
librosa
audioread

# Streamlit for alternative UI
streamlit

# Additional utilities for agents
psutil

pillow
mss
pycaw
comtypes
psutil
screen-brightness-control